pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stroke-risk-ml"
version = "1.0.0"
description = "End-to-end ML pipeline for brain stroke risk prediction with SHAP explainability, MLflow tracking, and FastAPI + Streamlit deployment."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Devendra Bhaginath Jadhav"},
]
keywords = ["machine-learning", "stroke-prediction", "healthcare", "mlops"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Healthcare Industry",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "imbalanced-learn>=0.11.0",
    "lightgbm>=4.1.0",
    "xgboost>=2.0.0",
    "optuna>=3.4.0",
    "shap>=0.43.0",
    "mlflow>=2.9.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "streamlit>=1.28.0",
    "plotly>=5.18.0",
    "seaborn>=0.13.0",
    "matplotlib>=3.8.0",
    "pyyaml>=6.0",
    "joblib>=1.3.0",
    "httpx>=0.25.0",
    "pydantic>=2.5.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.black]
line-length = 100
target-version = ["py310"]

